requirements.txt
Flask
Pillow
exiftool